+++ src/main/java/com/atlauncher/App.java
@@ -159,14 +159,6 @@ public class App {
     public static boolean disableErrorReporting = false;
 
     /**
-     * This allows skipping the hash checking when downloading files. It can be
-     * skipped with the below command line argument.
-     * <p/>
-     * --skip-hash-checking
-     */
-    public static boolean skipHashChecking = false;
-
-    /**
      * This forces the working directory for the launcher. It can be changed with
      * the below command line argument.
      * <p/>
@@ -363,19 +355,18 @@ public class App {
         boolean open = true;
 
         if (autoLaunch != null) {
-            if (InstanceManager.getInstances().stream()
-                    .anyMatch(instance -> instance.getSafeName().equalsIgnoreCase(autoLaunch))) {
-                Optional<Instance> instance = InstanceManager.getInstances().stream()
-                        .filter(i -> i.getSafeName().equalsIgnoreCase(autoLaunch)).findFirst();
-
-                if (instance.isPresent()) {
-                    LogManager.info("Opening Instance " + instance.get().launcher.name);
-                    if (instance.get().launch()) {
-                        open = false;
-                    } else {
-                        LogManager.error("Error Opening Instance " + instance.get().launcher.name);
-                    }
+            Optional<Instance> instance = InstanceManager.getInstances().stream().filter(
+                    i -> i.getName().equalsIgnoreCase(autoLaunch) || i.getSafeName().equalsIgnoreCase(autoLaunch))
+                    .findFirst();
+            if (instance.isPresent()) {
+                LogManager.info("Opening Instance " + instance.get().launcher.name);
+                if (instance.get().launch()) {
+                    open = false;
+                } else {
+                    LogManager.error("Error Opening Instance " + instance.get().launcher.name);
                 }
+            } else {
+                LogManager.error("Couldn't find instance with name of " + autoLaunch + " to auto launch.");
             }
         }
 
@@ -825,31 +816,56 @@ public class App {
     private static void parseCommandLineArguments(String[] args) {
         // Parse all the command line arguments
         OptionParser parser = new OptionParser();
-        parser.accepts("updated").withOptionalArg().ofType(Boolean.class);
-        parser.accepts("skip-setup-dialog").withOptionalArg().ofType(Boolean.class);
-        parser.accepts("skip-tray-integration").withOptionalArg().ofType(Boolean.class);
-        parser.accepts("disable-analytics").withOptionalArg().ofType(Boolean.class);
-        parser.accepts("disable-error-reporting").withOptionalArg().ofType(Boolean.class);
-        parser.accepts("skip-hash-checking").withOptionalArg().ofType(Boolean.class);
-        parser.accepts("force-offline-mode").withOptionalArg().ofType(Boolean.class);
-        parser.accepts("working-dir").withRequiredArg().ofType(String.class);
-        parser.accepts("base-launcher-domain").withRequiredArg().ofType(String.class);
-        parser.accepts("base-cdn-domain").withRequiredArg().ofType(String.class);
-        parser.accepts("base-cdn-path").withRequiredArg().ofType(String.class);
-        parser.accepts("allow-all-ssl-certs").withOptionalArg().ofType(Boolean.class);
-        parser.accepts("no-launcher-update").withOptionalArg().ofType(Boolean.class);
-        parser.accepts("no-console").withOptionalArg().ofType(Boolean.class);
-        parser.accepts("close-launcher").withOptionalArg().ofType(Boolean.class);
-        parser.accepts("debug").withOptionalArg().ofType(Boolean.class);
-        parser.accepts("debug-level").withRequiredArg().ofType(Integer.class);
-        parser.accepts("launch").withRequiredArg().ofType(String.class);
-        parser.accepts("proxy-type").withRequiredArg().ofType(String.class);
-        parser.accepts("proxy-host").withRequiredArg().ofType(String.class);
-        parser.accepts("proxy-port").withRequiredArg().ofType(Integer.class);
-        parser.accepts("config-override").withRequiredArg().ofType(String.class);
+        parser.accepts("updated", "If the launcher was just updated.").withOptionalArg().ofType(Boolean.class);
+        parser.accepts("skip-setup-dialog",
+                "If the first time setup dialog should be skipped, using the defaults. Note that this will enable analytics by default.")
+                .withOptionalArg().ofType(Boolean.class);
+        parser.accepts("skip-tray-integration", "If the tray icon should not be enabled.").withOptionalArg()
+                .ofType(Boolean.class);
+        parser.accepts("disable-analytics", "If analytics should be disabled.").withOptionalArg().ofType(Boolean.class);
+        parser.accepts("disable-error-reporting", "If error reporting should be disabled.").withOptionalArg()
+                .ofType(Boolean.class);
+        parser.accepts("working-dir", "This forces the working directory for the launcher.").withRequiredArg()
+                .ofType(String.class);
+        parser.accepts("base-launcher-domain", "The base launcher domain.").withRequiredArg().ofType(String.class);
+        parser.accepts("base-cdn-domain", "The base CDN domain.").withRequiredArg().ofType(String.class);
+        parser.accepts("base-cdn-path", "The path on the CDN used for downloading files.").withRequiredArg()
+                .ofType(String.class);
+        parser.accepts("allow-all-ssl-certs",
+                "This will tell the launcher to allow all SSL certs regardless of validity. This is insecure and only intended for development purposes.")
+                .withOptionalArg().ofType(Boolean.class);
+        parser.accepts("no-launcher-update",
+                "This forces the launcher to not check for a launcher update. It can be enabled with the below command line argument.")
+                .withOptionalArg().ofType(Boolean.class);
+        parser.accepts("no-console", "If the console shouldn't be shown.").withOptionalArg().ofType(Boolean.class);
+        parser.accepts("close-launcher", "If the launcher should be closed after launching an instance.")
+                .withOptionalArg().ofType(Boolean.class);
+        parser.accepts("debug", "If debug logging should be enabled.").withOptionalArg().ofType(Boolean.class);
+        parser.accepts("debug-level", "The level of debug logging that should be logged.").withRequiredArg()
+                .ofType(Integer.class);
+        parser.accepts("launch",
+                "The name of an instance to automatically launch. Can be the instances directory name in the file system or the full name of the instance.")
+                .withRequiredArg().ofType(String.class);
+        parser.accepts("proxy-type", "The type of proxy to use. Can be \"SOCKS\", \"DIRECT\" or \"HTTP\".")
+                .withRequiredArg().ofType(String.class);
+        parser.accepts("proxy-host", "The host of the proxy to use.").withRequiredArg().ofType(String.class);
+        parser.accepts("proxy-port", "The port of the proxy to use.").withRequiredArg().ofType(Integer.class);
+        parser.accepts("config-override", "A JSON string to override the launchers config.").withRequiredArg()
+                .ofType(String.class);
+        parser.acceptsAll(Arrays.asList("help", "?"), "Shows help for the arguments for the application.").forHelp();
+
+        try {
+            parser.printHelpOn(System.out);
+        } catch (IOException e1) {
+            e1.printStackTrace();
+        }
 
         OptionSet options = parser.parse(args);
         autoLaunch = options.has("launch") ? (String) options.valueOf("launch") : null;
+
+        if (options.has("help")) {
+            System.exit(0);
+        }
 
         if (options.has("updated")) {
             wasUpdated = true;
@@ -930,11 +946,6 @@ public class App {
         closeLauncher = options.has("close-launcher");
         if (closeLauncher) {
             LogManager.debug("Closing launcher once Minecraft is launched!");
-        }
-
-        skipHashChecking = options.has("skip-hash-checking");
-        if (skipHashChecking) {
-            LogManager.debug("Skipping hash checking! Don't ask for support with this enabled!");
         }
 
         if (options.has("proxy-type") && options.has("proxy-host") && options.has("proxy-port")) {